refactor: move middleware to dedicated module and enhance CORS handling
- Move NoCacheMiddleware to new middleware.py module
- Create LocalhostCORSMiddleware that allows any localhost/127.0.0.1 request
- Update listen.py to use new middleware
